Bring the indexing constants back.
For motors, we need independent PWM ports.  For everything else, we still need
the indexing constants.  I've also changed the ports so that they are keyed
off of the value of the constants: the drive PWM port for a given wheel will
always be equal to its indexing constant and the pivot PWM port for a given
wheel will always be the drive port + 4.