Add theoretical support for keyboard input via NetworkTables.
You'll need to have NetworkTables-input.jar running, of course.

We don't have a NetworkButton class anymore (at least as of this time -- not
locally -- even through the docs are online.)  So I'm assuming that
NetworkButtons just pluck a boolean out of the NetworkTable and I wrote code
to do the same thing.