Make the shooter command automatically self-cancel if the trigger is released during the wait period.
Without this change, pressing the trigger and then releasing it before
0.4 seconds (the current wait period) have passed causes the belts to
continue to go off as if they had a mind of their own.

What's really happening is that we *did* stop the belt motors...only
to have them start again when the wait period was done and the rest of
the old, non-canceled command executed!  The solution is to make sure
the old command gets canceled upon release.

I don't think we need to worry about this under ordinary circumstances--
only when WaitCommands are involved.