Attempt to fix the rollover wraparound problem by avoiding setReference().
PIDController.setReference() has an unfortunate tendency to cause the
motor to spin towrad its goal without regards to which spinning direction
would be the most efficient.  For instance, in an extreme example, if
the current position was at 1 degree and we asked it to spin to 359
degrees, it would spin 358 degrees, wrapping all the way around, instead
of just going 2 degrees in the other direction.

This is an attempt to fix that.  Instead of calling setReference(), we
continuously get the encoder's position, determine the shortest direction
ourselves, and spin towards it until we reach a deadzone threshold.

Does it work?  Not sure.  It's totally untested.  We're waiting for the
electronics board to come back together first -- that's why these
changes are squirreled away in a branch.