Wheels calibrate when test button is pressed.
When the test button is pressed, the pivot motors reset the wheels to the initial position of the wheels after the initial movement caused by enabling.

We are not using fudge factors anymore.

TODO: Allow the pivot motors to reset the wheels when the robot is disabled.
Prevent the initial movemnt caused by enabling.