Removed all calibrate function calls.
We observed odd behavior using the calibrate function. Upon commenting it out, the drive appears to calibrate itself without any command to do so.

Additionally, we observed that the Neo encoders have a zero value on power-up, so we assume them to be relative (instead of absolute).

We still need to physically calibrate the wheels before power-up so that the drive remembers its inital position.